+++ MANIFEST.in
@@ -1,3 +1,4 @@
 include *.md
 include *.rst
+recursive-include icclim *.json
 recursive-include icclim *.so
